requirements.txt
aio-binance-library
asyncpg
flake8
flake8-bugbear
pep8-naming